modal refacto + ui + cairo
