+++ apps/mobile/src/modules/Post/styles.ts
@@ -46,10 +46,18 @@ export default ThemedStyleSheet((theme) => ({
     marginBottom: Spacing.medium,
     color: theme.colors.text,
   },
+  // contentImage: {
+  //   width: '100%',
+  //   height: 'auto',
+  //   resizeMode: 'cover',
+  //   borderRadius: 8,
+  //   overflow: 'hidden',
+  //   marginTop: Spacing.small,
+  // },
   contentImage: {
     width: '100%',
-    height: 'auto',
-    resizeMode: 'cover',
+    height: '100%',
+    // resizeMode: 'cover',
     borderRadius: 8,
     overflow: 'hidden',
     marginTop: Spacing.small,
