+++ apps/mobile/src/screens/Feed/styles.ts
@@ -23,6 +23,8 @@ export default ThemedStyleSheet((theme) => ({
     position: 'absolute',
     bottom: Spacing.large,
     right: Spacing.pagePadding,
+    // backgroundColor:theme.colors.primary,
+    color:theme.colors.primary
   },
 
   stories: {
