+++ packages/afk_nostr_sdk/src/utils/relay.ts
@@ -1,4 +1,5 @@
 export const RELAYS_PROD = ['wss://relay.n057r.club', 'wss://relay.nostr.net', "wss://relay.primal.net", "wss://relay.nostr.band", "wss://purplepag.es"];
+export const RELAYS_TEST= ['wss://relay.n057r.club', 'wss://relay.nostr.net'];
 
 export const AFK_RELAYS =
   process.env.EXPO_NODE_ENV == 'production' || process.env.NODE_ENV == 'production'
@@ -13,7 +14,7 @@ export const AFK_RELAYS =
       // 'ws://nostr-relay-nestjs-production.up.railway.app',
       // 'wss://nostr-relay-nestjs-production.up.railway.app',
       'ws://localhost:8080', // comment if you don't run a relayer in localhost
-      ...RELAYS_PROD
+      ...RELAYS_TEST
     ];
 
 // export const AFK_RELAYS = [
